Rename(db_handler): rename db_manager to db_handler
